+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -101,5 +101,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-09-29T03:53:49.945Z")
+    created: Optional[datetime] = parse_datetime("2025-09-29T07:35:50.555Z")
     r"""The date and time the resource was created"""