+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -603,7 +603,7 @@ class RunAgentRequestToolHTTPTool(BaseModel):
     http: RunAgentRequestToolHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K69SP322N8HZK6Y40VNVN53J"
+        "01K6A6CKCZ5RZA1T4FPSK916V2"
     )
 
     requires_approval: Optional[bool] = False