+++ packages/orq-rc/src/orq_ai_sdk/models/updatecontactop.py
@@ -181,7 +181,7 @@ class UpdateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-29T03:53:46.885Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-29T07:35:47.618Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")